Corrected issues when Redacting was enabled and comprehend redaction was disabled
