+++ lambda/qnabot-common-layer/qnabot/utilities.js
@@ -1,0 +1,15 @@
+function isJson(str) {
+    if(!str){
+        return false
+    }
+    try {
+        JSON.parse(str);
+    } catch (e) {
+        return false;
+    }
+    return true;
+}
+
+module.exports = {
+    isJson: isJson
+}
